Rename AlphaCalc => NimFieldCalculator, FieldTable => NimFieldConstants, and split out ImpartialTermAlgebra into a separate file
