+++ lib/core/src/main/scala/org/cgsuite/core/impartial/arithmetic/NimFieldCalculator.scala
@@ -1,22 +1,38 @@
 package org.cgsuite.core.impartial.arithmetic
 
-import java.util
-
+import better.files._
 import ch.qos.logback.classic.Logger
+import ch.qos.logback.classic.spi.ILoggingEvent
+import ch.qos.logback.core.{ConsoleAppender, OutputStreamAppender}
 import com.typesafe.scalalogging.LazyLogging
+import org.cgsuite.core.impartial.arithmetic.ImpartialTermAlgebra.Element
 import org.cgsuite.core.{GeneralizedOrdinal, Integer, SmallInteger, Values}
 import org.cgsuite.dsl._
 import org.slf4j.LoggerFactory
-import better.files._
-import ch.qos.logback.classic.spi.ILoggingEvent
-import ch.qos.logback.core.{ConsoleAppender, OutputStreamAppender}
 
 import scala.collection.mutable
 
-object AlphaCalc extends LazyLogging {
+/* A tool for calculating Lenstra's constants Q(p) and excess(p) for odd primes p.
+ *
+ * Multiplication in transfinite nim arithmetic depends on the values of Q(p) and
+ * excess(p). Calculating those values is computationally expensive, so in CGSuite's
+ * implementation of nim arithmetic, they are defined as constants
+ * (cf. NimFieldConstants.scala). This tool can be used to recompute them - thus
+ * documenting their provenance - or, given sufficient compute resources, to extend
+ * the list.
+ *
+ * The tool is essentially a separate implementation of transfinite nim arithmetic,
+ * specifically optimized for computing Q(p) and excess(p). The strategy is to
+ * construct a complete materialized representation of the minimal finite field
+ * containing kappa_{f(p)} (Lenstra's notation) and carry out the requisite operations
+ * within it.
+ *
+ * The notation and approach follow Lenstra (1977), On the algebraic closure of two.
+ */
+object NimFieldCalculator extends LazyLogging {
 
   def main(args: Array[String]): Unit = {
-    run(from = 263, preloadTo = 1009)
+    run()
   }
 
   private val excessCache = mutable.Map[Int, Int]()
@@ -31,15 +47,15 @@ object AlphaCalc extends LazyLogging {
     appender.setEncoder(consoleAppender.asInstanceOf[ConsoleAppender[ILoggingEvent]].getEncoder)
     appender.start()
     rootLogger.addAppender(appender)
-    val kPreload = FieldTable.primes.indexOf(preloadTo)
-    FieldTable.excess.indices.slice(1, kPreload + 1) foreach { k => excessCache.put(FieldTable.primes(k), FieldTable.excess(k)) }
-    FieldTable.qSet.indices.slice(1, kPreload + 1) foreach { k => qSetCache.put(FieldTable.primes(k), FieldTable.qSet(k)) }
+    val kPreload = NimFieldConstants.primes.indexOf(preloadTo)
+    NimFieldConstants.excess.indices.slice(1, kPreload + 1) foreach { k => excessCache.put(NimFieldConstants.primes(k), NimFieldConstants.excess(k)) }
+    NimFieldConstants.qSet.indices.slice(1, kPreload + 1) foreach { k => qSetCache.put(NimFieldConstants.primes(k), NimFieldConstants.qSet(k)) }
     println()
     println(f"${"p"}%4s ${"f(p)"}%4s ${"Q(f(p))"}%15s ${"exponent"}%7s excess ${"alpha_p"}%20s ${"t(sec)"}%8s     alpha_seq")
-    var k = FieldTable.primes.indexOf(from)
+    var k = NimFieldConstants.primes.indexOf(from)
     while (true) {
       val startTime = System.currentTimeMillis()
-      val p = FieldTable.primes(k)
+      val p = NimFieldConstants.primes(k)
       print(f"$p%4d ${f(p)}%4d")
       val qSet = this.qSet(p)
       val qSetStr = s"{${qSet mkString ","}}"
@@ -48,7 +64,7 @@ object AlphaCalc extends LazyLogging {
       val excess = this.excess(p)
       print(f" $excess%6d ${alphap(p)}%20s")
       val excessSeq = (1 to k) map { i =>
-        this.excessCache.getOrElse(FieldTable.primes(i), "??")
+        this.excessCache.getOrElse(NimFieldConstants.primes(i), "??")
       } mkString ","
       val durationSecs = (System.currentTimeMillis() - startTime) / 1000.0
       println(f" $durationSecs%8.1f     $excessSeq")
@@ -75,6 +91,7 @@ object AlphaCalc extends LazyLogging {
     result
   }
 
+  // true if   2^pow === 1 (mod n).
   def dividesTwoPowerMinus1(n: Int, pow: Int): Boolean = {
     var twoPower = 1
     for (_ <- 1 to pow) {
@@ -211,7 +228,7 @@ object AlphaCalc extends LazyLogging {
   def kappa(h: Int): IndexedSeq[Int] = {
 
     // Find the smallest prime divisor of n
-    val p = FieldTable.primes.find { h % _ == 0 } getOrElse {
+    val p = NimFieldConstants.primes.find { h % _ == 0 } getOrElse {
       throw new IllegalArgumentException(h.toString)
     }
 
@@ -263,7 +280,7 @@ object AlphaCalc extends LazyLogging {
     if (p == 2) {
       Values.two.exp(Values.two.exp(Integer(n - 1)).asInstanceOf[Integer]).asInstanceOf[Integer]
     } else {
-      val k = FieldTable.primes.indexOf(p)
+      val k = NimFieldConstants.primes.indexOf(p)
       // Value is omega^(omega^(k-1) * p^(n-1))
       val exponent = Integer(k - 1).omegaPower.asInstanceOf[GeneralizedOrdinal] * Integer(p).exp(Integer(n - 1)).asInstanceOf[GeneralizedOrdinal]
       exponent.omegaPower.asInstanceOf[GeneralizedOrdinal]
@@ -276,7 +293,7 @@ object AlphaCalc extends LazyLogging {
     if (q <= 1) {
       throw new IllegalArgumentException(s"Not a prime power: $q")
     }
-    val p = FieldTable.primes.find(q % _ == 0).getOrElse {
+    val p = NimFieldConstants.primes.find(q % _ == 0).getOrElse {
       throw new IllegalArgumentException(s"Base out of range: $q")
     }
     var n = 1
@@ -289,206 +306,6 @@ object AlphaCalc extends LazyLogging {
       throw new IllegalArgumentException(s"Not a prime power: $q")
     }
     (p, n)
-  }
-
-}
-
-object ImpartialTermAlgebra {
-
-  def apply(qComponents: IndexedSeq[Int]): ImpartialTermAlgebra = {
-    new ImpartialTermAlgebra(qComponents.distinct.sortBy(AlphaCalc.primePower).toArray)
-  }
-
-}
-
-class ImpartialTermAlgebra private (val qComponents: Array[Int]) extends LazyLogging {
-
-  // TODO Validate that qComponents is closed under q/p
-
-  val qDegrees = qComponents map { AlphaCalc.primePower(_)._1 }
-  val degreeProducts = (0 to qDegrees.length).map { qDegrees.take(_).product }.toArray
-  val termCount = degreeProducts.last
-  val accumulator = new Array[Long]((termCount + 63) >> 6)
-
-  // Associates to each component kappa_q the element given by
-  // kappa_q^p (where p is the prime divisor of q).
-  val kappaTable: IndexedSeq[Element] = qComponents.indices map { index =>
-    (qComponents(index), qDegrees(index)) match {
-      case (_, 2) => Element(Seq(degreeProducts(index), degreeProducts(index) - 1))
-      case (q, p) if q == p =>
-        val qSet = AlphaCalc.qSet(p)
-        val excess = AlphaCalc.excess(p)
-        assert(excess == 0 || excess == 1 || SmallInteger.isTwoPower(excess))
-        val kappaBlocks = qSet.map { q2 =>
-          val index = qComponents.indexOf(q2)
-          assert(index >= 0, s"$q2 not in ${qComponents.toSeq}")
-          degreeProducts(index)
-        }
-        Element(excess match {
-          case 0 => kappaBlocks
-          case _ => kappaBlocks :+ SmallInteger.lb(excess)
-        })
-      case _ => Element(Seq(degreeProducts(index - 1)))
-    }
-  }
-
-  logger.info {
-    val ktStrings = kappaTable map { element => s"{${element.terms.sorted mkString ","}}" }
-    s"(q, kappa_q^p)   =   ${qComponents zip ktStrings map { case (q, ktString) => s"$q -> $ktString"} mkString "   "}"
-  }
-
-  //for (index <- qComponents.indices) { println(s"${qComponents(index)}: ${kappaTable(index) mkString ","}") }
-
-  val qPowerTimesTermTable = new Array[Array[Array[Element]]](qComponents.length)
-  qPowerTimesTermTable.indices foreach { index =>
-    qPowerTimesTermTable(index) = new Array[Array[Element]](qDegrees(index))
-    1 until qDegrees(index) foreach { qExponent =>
-      qPowerTimesTermTable(index)(qExponent) = new Array[Element](termCount)
-    }
-  }
-
-  for (qIndex <- qComponents.indices;
-       qExponent <- 1 until qDegrees(qIndex);
-       term <- 0 until termCount) {
-    qPowerTimesTerm(qIndex, qExponent, term)
-  }
-
-  def qPowerTimesTerm(qIndex: Int, qExponent: Int, term: Int): Element = {
-
-    val cached = qPowerTimesTermTable(qIndex)(qExponent)(term)
-    if (cached != null) {
-      cached
-    } else {
-      val result = qPowerTimesTermCalc(qIndex, qExponent, term)
-      qPowerTimesTermTable(qIndex)(qExponent)(term) = result
-      result
-    }
-
-  }
-
-  def qPowerTimesTermCalc(qIndex: Int, qExponent: Int, term: Int): Element = {
-
-    val p = qDegrees(qIndex)
-    val qExponentInTerm = (term % degreeProducts(qIndex + 1)) / degreeProducts(qIndex)
-    val qExponentNew = qExponent + qExponentInTerm
-
-    if (qExponentNew < p) {
-      Element(Seq(term + qExponent * degreeProducts(qIndex)))
-    } else {
-
-      assert(qExponentNew < 2 * p)
-      val highOrderPart: Int = (term / degreeProducts(qIndex + 1)) * degreeProducts(qIndex + 1) + (qExponentNew % p) * degreeProducts(qIndex)
-      val lowOrderPart: Int = term % degreeProducts(qIndex)
-      logger.debug("parts = " + highOrderPart + "," + lowOrderPart)
-      // We want to return highOrderPart * (lowOrderPart * kappa^p).
-      // Retrieve the summands of kappa^p.
-      val kappaExpansion = kappaTable(qIndex)
-      logger.debug("kexp = " + kappaExpansion)
-      // Compute the summands of lowOrderPart * kappa^p.
-      val terms = mutable.Set[Int]()
-      for (term <- kappaExpansion.terms) {
-        for (productTerm <- termTimesTerm(lowOrderPart, term).terms) {
-          if (terms.contains(productTerm))
-            terms.remove(productTerm)
-          else
-            terms.add(productTerm)
-        }
-      }
-      Element(terms.toSeq.sorted map { highOrderPart + _ })
-
-    }
-
-  }
-
-  def termTimesTerm(x: Int, y: Int): Element = {
-    assert(x >= 0)
-    assert(y >= 0)
-    var terms = mutable.Set(y)
-    for (xIndex <- qComponents.length - 1 to 0 by -1) {
-      val xExponent = (x % degreeProducts(xIndex + 1)) / degreeProducts(xIndex)
-      if (xExponent > 0) {
-        val newTerms = mutable.Set[Int]()
-        for (term <- terms) {
-          for (productTerm <- qPowerTimesTerm(xIndex, xExponent, term).terms) {
-            if (newTerms.contains(productTerm)) newTerms.remove(productTerm)
-            else newTerms.add(productTerm)
-          }
-        }
-        terms = newTerms
-      }
-    }
-    Element(terms.toSeq.sorted)
-  }
-
-  @inline
-  private def flipAccumulatorTerm(x: Int): Unit = {
-    accumulator(x / 64) ^= 1L << x
-  }
-
-  private def accumulatorContains(x: Int): Boolean = {
-    (accumulator(x / 64) & (1L << x)) != 0
-  }
-
-  def accumulateTermProduct(x: Int, y: Int): Unit = {
-    if (y == 0) {
-      flipAccumulatorTerm(x)
-    } else {
-      // Take the highest-order remaining qComponent from y.
-      var index = qComponents.length
-      while (y / degreeProducts(index) == 0) {
-        index -= 1
-      }
-      // Multiply that component by x to get a sum of terms.
-      val product = qPowerTimesTerm(index, y / degreeProducts(index), x)
-      // Apply each component of the sum.
-      for (term <- product.terms) {
-        accumulateTermProduct(term, y % degreeProducts(index))
-      }
-    }
-  }
-
-  def multiply(a: Element, b: Element): Element = {
-    util.Arrays.fill(accumulator, 0)
-    for (x <- a.terms; y <- b.terms) {
-      accumulateTermProduct(x, y)
-    }
-    Element(0 until termCount collect {
-      case x if accumulatorContains(x) => x
-    })
-  }
-
-  def pow(x: Element, n: BigInt): Element = {
-    var curpow: Element = x
-    var result: Element = Element(Seq(0))
-    var i = n
-    var checkpointTime = System.currentTimeMillis()
-    while (i != BigInt(0)) {
-      if (i.testBit(0)) {
-        result = multiply(result, curpow)
-      }
-      curpow = multiply(curpow, curpow)
-      if (System.currentTimeMillis() - checkpointTime >= 60000L) {
-        val bitsComplete = n.bitLength - i.bitLength
-        logger.info(f"$bitsComplete/${n.bitLength} bits complete (${bitsComplete.toDouble/n.bitLength}%2.1f%%); curpow/result has ${curpow.terms.size}/${result.terms.size} terms")
-        checkpointTime = System.currentTimeMillis()
-      }
-      i >>= 1
-    }
-    result
-  }
-
-}
-
-object Element {
-
-  def apply(terms: Seq[Int]): Element = new Element(terms.sorted)
-
-}
-
-case class Element private (terms: Seq[Int]) {
-
-  def isOne: Boolean = {
-    terms.size == 1 && terms.head == 0
   }
 
 }
